requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
Pypdf
chromadb